Disable sanitizers in tests when using any kind of instrumentations with MSVC.
Clang uses a massive amount of memory (16GB) in one of the source files when address sanitizer is in use. This is due to how SEH exception handling works in MSVC. LLVM generates so called cleanupret instructions for each try catch, which might transfer control back to the caller. Before that happens it needs to reset shadow map state generated by ASAN. A normal function generated by clang has a single point of return which even in a large function would cause all the code of the state reset to be emitted there and nowhere else. SEH exception handling however inserts a cleanupret instruction per try catch. For every try catch in the function, the state reset has to generated, causing incredibly code duplication and therefore memory consumption in the compiler.