Fix lifetime of TypedefInfo objects.
I was wrongly under the impression that the lifetime of these would not be invalidated by the scope but apparently, Scope objects are not getting moved inside of the std::vector, or possibly tsl::ordered_map just does not give any guarantees about reference stability. It does not give iterator stability guarantee for operator=, so I am going to assume it's the same for reference stability. Therefore I opted to use a deque to simply contain all TypedefInfos, similar to how I do for Structs, Unions and Enums.