Rewrote the way attributes are parsed and applied.
They are now done in two stages, 1) parsing the attribute 2) applying the attribute to the applicant. The first uses information gathered by deconstructing it's tag type which may contain fields that have to optionally be populated. Afterwards the ParsedAttribute can be applied onto an applicant. This makes it a lot easier to add a new attribute and also allows code in the semantics stage to check whether an attribute exists before even applying it. Very useful for cases such as dllimport which implies extern as an example