Changed attributes of the *Declaration classes to be mutable as subsequent and/or previous declarations may change these.
In the process a bug that made attributes disappear was found which then revealed another bug inside of the gnu_inline attribute which has also been fixed.

This current approach seems to be better but maybe in the future I should simply look up a previous existence of the symbol, do merging of attributes and only then construct the new Function/Variable. Something to think about