Optimize happy path of primitiveTypeSpecifiersToType
Unless an error occurs, there is really no need to keep a text presentation of the type specifiers, side by side as we are constructing the primitive type. Therefore, simply construct the text representation right before emitting an error. This leads to the happy path actually having 0 allocations in this function